skincare-store-main/backend/requirements.txt
Django==4.2
djangorestframework==3.15.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
django-cors-headers==3.14.0
Pillow==9.5.0
